[REF] product_pricelist_direct_print_xlsx
Code refactoring made with purpose of making the module easier
to extend. Since `xlsxwriter` is "write only" library, there was no way
to easily change the properties of printed table. The changes are
focused on separating preparation of data / labels / formats.

That way any extension module can change the order of the columns, data
formats or add new columns anywhere in the table. For example, prior
to this change, to add a new column before the "Description" column,
only way was to override `_create_product_pricelist_sheet` and
`_fill_data` methods or rewrite  whole workbook from scratch after the
original methods are done.